pyproject.toml
[tool.poetry]
name = "monty-assignment"
version = "0.1.0"
description = ""
authors = ["Emile Ibrahim <emileib.lb@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
motor = "^3.5.0"
pydantic-settings = "^2.3.4"
beanie = "^1.26.0"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiofiles = "^24.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
